added frontend

+++ backend/src/services/abstract-room.service.ts
@@ -26,11 +26,12 @@ export function abstractCreateOrJoinRoom(
     if (userLateRemoveTimers[userId]) {
         console.log(`Reconnect user ${userName} (${userId}) to room ${roomId}`);
         roomRepository.setUserActivity(roomId, userId, true);
-        clearTimeout(userLateRemoveTimers[userId]);
     } else {
         console.log(`Add user ${userName} (${userId}) to room ${roomId}`);
         roomRepository.addUser(roomId, { id: userId, name: userName, active: true });
     }
+    clearTimeout(userLateRemoveTimers[userId]);
+    delete userLateRemoveTimers[userId];
     roomUserWsMappingService.upsertUser(roomId, userId, ws);
 
     wsSubscription(ws, roomId, userId, userActionCallback);
@@ -42,33 +43,43 @@ export function abstractCreateOrJoinRoom(
 }
 
 function wsSubscription(ws: WS, roomId: RoomId, userId: UserId, userActionCallback: UserActionStrategy) {
-    const userName = roomRepository.getRoom(userId).users[userId].name;
+    const userName = roomRepository.getRoom(roomId).users[userId].name;
     console.log(`Subscribe on user actions for ${userName} (${userId}) in room ${roomId}`);
     ws.on("message", (msg: string): void => {
         if (msg === "H") {
             //console.log(`client H for ${userName} (${userId})`);
         } else {
+            //TODO: app crushed on JSON.parse('test')
             const userAction = JSON.parse(msg);
             userActionCallback(roomId, userId, userAction);
         }
         broadcastRoom(roomId);
     });
-    ws.on("close", () => {
+    ws.on("close", (a, b) => {
         console.log(`WS for user (${userName} : ${userId}) in room (${roomId}) was closed`);
+        console.log(a, b);
         roomUserWsMappingService.removeUser(roomId, userId);
         roomRepository.setUserActivity(roomId, userId, false);
         userLateRemoveTimers[userId] = setTimeout(() => {
+            if (!userLateRemoveTimers[userId]) {
+                console.warn(
+                    `Timeout for userId was cleared so skip deletion user ${userName} (${userId}) from room ${roomId}`,
+                );
+                return;
+            }
+
             console.info(`Lazy remove for user (${userName} : ${userId}) in room (${roomId})`);
             roomRepository.removeUser(roomId, userId);
             delete userLateRemoveTimers[userId];
 
             broadcastRoom(roomId);
 
-            if (isEmptyRoom(roomId)) {
-                console.log(`Room ${roomId} is empty so remove it`);
-                roomRepository.removeRoom(roomId);
-                roomUserWsMappingService.removeRoom(roomId);
-            }
+            // will remove in startCleanupOldRooms
+            // if (isEmptyRoom(roomId)) {
+            //     console.log(`Room ${roomId} is empty so remove it`);
+            //     roomRepository.removeRoom(roomId);
+            //     roomUserWsMappingService.removeRoom(roomId);
+            // }
         }, LAZY_REMOVE_TIMEOUT);
     });
     ws.on("error", (e: any) => {
@@ -84,7 +95,7 @@ const broadcastRoom = (roomId: RoomId): void => {
     Object.values(userWsMap).forEach((ws: WS) => ws.send(JSON.stringify(room)));
 };
 
-const isEmptyRoom = (roomId: RoomId): boolean => {
-    const room = roomRepository.getRoom(roomId);
-    return !Object.keys(room.users).length;
-};
+// const isEmptyRoom = (roomId: RoomId): boolean => {
+//     const room = roomRepository.getRoom(roomId);
+//     return !Object.keys(room.users).length;
+// };
